+++ tests/integrationtests/api/conftest.py
@@ -84,11 +84,13 @@ def metadata_report_multilevel_mapping_saturation():
 
 
 @pytest.fixture
-def metadata_quality_dimension_completeness():
+def metadata_quality_dimension():
     return {
-        "completeness": {
-            "name": "Completeness",
-            "description": "something that is still a TODO",
+        "minimal": {
+            "name": "Minimal",
+            "description": "A minimal quality dimension"
+            " definition for testing purposes.",
+            "source": None,
         }
     }